requirements.txt
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
matplotlib>=3.8.0
plotly>=5.20.0
pyyaml>=6.0.1
optuna>=3.6.0
reportlab>=4.1.0
weasyprint>=62.0
pytest>=8.1.0
